+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to
   and `EthernetFrame`.
 * Added `driver_set_link_speed(speed)` and `driver_set_link_full_duplex(flag)`.
 * Added `EthernetClass::renewDHCP()`.
+* Added `EthernetClass::interfaceName()`.
 
 ### Changed
 * Separated setting the MAC address from driver initialization.